# api/openapi.yaml
openapi: 3.1.0
info:
  title: GuardPipe API
  version: "1.1"
  summary: Software supply-chain security scanning API.
  description: |
    GuardPipe scans a project across seven SDLC stages and produces a single
    explainable 0-100 risk score with AI-authored explanations and patches.

    This document is machine-derived from the currently implemented HTTP
    surface (`internal/transport/http/router.go` and its handlers/DTOs), not
    hand-aspirational — every path, parameter, and schema here corresponds to
    code that actually runs today. The **human-authored source of truth**
    for the intended, eventual API is
    [`documentation/07-api-specification.md`](../documentation/07-api-specification.md)
    (see its §12); per that document's own rule, "if code and docs disagree,
    one of them is a bug — open an issue, don't let them silently diverge."
    A few endpoints described there (finding detail/triage/history/explain,
    the dashboard/trend endpoints, SSE progress, GitHub webhooks) are not yet
    wired into the router and are intentionally absent from this file until
    they are. `GET /scans/{id}/export` *is* wired (json/csv/pdf, pulled
    forward ahead of the rest of Phase 13's `reporting` module) and appears
    below.

    ### Conventions
    - Base path `/api/v1` for the product API; `/healthz`, `/readyz`,
      `/version` are unversioned infrastructure endpoints.
    - All JSON fields are `snake_case`. IDs are UUIDv4 strings. Timestamps
...
      below).
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/projects/{id}/scans:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      tags: [Scans]
      summary: Start a scan
      description: >-
        Asynchronous by design — returns 202 immediately and the client
        polls `GET /scans/{id}/progress`; a scan takes minutes, so holding
        the HTTP connection open would be wrong at every level. Role
        required member or above. Rate-limited to 10/hour per user (per
        `documentation/07-api-specification.md` §1.6; not yet
        independently rate-limited in code beyond the generic authenticated
        limit — see that document's table).
      operationId: createScan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScanRequest"
      responses:
        "202":
          description: Scan queued.
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
        "400":
          description: >-
            Malformed request, or `scan.no_engines_selected` when `type` is
            `partial` with an empty `engines` list.
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: A scan is already running for this project (`scan.already_running`).
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/ProblemDetails" }
    get:
      tags: [Scans]
      summary: Scan history for a project
      operationId: listProjectScans
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of the project's scans, newest first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans:
    get:
      tags: [Scans]
      summary: Scan history across every project in the organisation
      operationId: listOrgScans
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of the organisation's scans, newest first, each row carrying its project's name.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgScanListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/scans/{id}:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: Scan detail
      description: Full detail including per-engine job status.
      operationId: getScan
      responses:
        "200":
          description: Scan detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans/{id}/progress:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: Lightweight scan progress
      description: >-
        Deliberately tiny response, meant to be polled every ~2 seconds by
        the UI while a scan runs.
      operationId: getScanProgress
      responses:
        "200":
          description: Current progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    post:
      tags: [Scans]
      summary: Cancel a scan
      description: Role required member or above.
      operationId: cancelScan
      responses:
        "204":
          description: Cancellation requested.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans/{id}/findings:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Findings]
      summary: List a scan's findings
      description: >-
        Pagination only in the current implementation — the severity/
        engine/status/cwe/file_path/q/sort filters described in
        `documentation/07-api-specification.md` §1.5 are Phase 13 work and
        not yet wired into this handler.
      operationId: listFindings
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of the scan's findings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FindingListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scans/{id}/export:
    parameters:
      - $ref: "#/components/parameters/ScanId"
    get:
      tags: [Scans]
      summary: Export a scan report
      description: >-
        A self-contained report snapshot (metadata, every job's stats —
        including pentest's coverage summary — every finding with its
        deterministic remediation, and an AI-authored executive summary when
        AI is configured), distinct from `GET /scans/{id}`'s live shape.
        `modules/reporting`'s export slice only, pulled forward ahead of the
        rest of Phase 13 — no `risk` block, since `modules/scoring` isn't
        built yet. `format=sarif` is not implemented and returns 400, not
        the 501 an earlier draft of this endpoint described.
      operationId: exportScan
      parameters:
        - name: format
          in: query
          description: Defaults to `json` when omitted.
          schema:
            type: string
            enum: [json, csv, pdf]
      responses:
        "200":
          description: >-
            The report in the requested format, served as an attachment
            (`Content-Disposition: attachment`).
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ------------------------------------------------------------------ Rules
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        `Authorization: Bearer <access_token>`. The access token is short-lived
        (`expires_in` seconds from login/refresh); refresh via
        `POST /auth/refresh` using the `gp_refresh` cookie.
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Project ID.
    ScanId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
      description: Scan ID.
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: 1-based page number.
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
      description: Items per page, clamped to 100.
  responses:
    Validation:
      description: One or more fields are invalid.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    Unauthorized:
      description: Missing, invalid, or expired access token.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    Forbidden:
      description: Authenticated but the caller's role does not permit this action.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    NotFound:
      description: Not found, or not owned by the caller's organisation.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
  schemas:
...
    FieldError:
      type: object
      properties:
        field: { type: string }
        message: { type: string }
      required: [field, message]
    ProblemDetails:
      type: object
      description: RFC 9457 problem-details error body.
      properties:
        type:
          type: string
          format: uri
          description: Stable identifier URI for this error kind, namespaced under `https://guardpipe.dev/errors/`.
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
          description: Safe for display to a user. Never a stack trace, SQL, or an internal path.
        instance:
          type: string
          description: The request path.
        code:
          type: string
          description: Machine-readable, stable, namespaced. This is what the frontend switches on — never `title` or `detail`.
        request_id:
          type: string
          description: Matches the corresponding log entry.
        errors:
          type: array
          items: { $ref: "#/components/schemas/FieldError" }
          description: Present only on validation failures.
      required: [type, title, status, detail, code]
    Pagination:
      type: object
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }
      required: [page, page_size, total, total_pages]

    # --------------------------------------------------------------- Enums
...
    Severity:
      type: string
      enum: [critical, high, medium, low, informational]
    Confidence:
      type: string
      enum: [high, medium, low]
    FindingStatus:
      type: string
      enum: [open, acknowledged, suppressed, fixed, false_positive]
    EngineID:
      type: string
      enum: [docreview, codescan, depscan, containerscan, k8sscan, cicdscan, pentest]
...
    FindingSource:
      type: string
      enum: [rule, ai]
...
    ScanType:
      type: string
      enum: [full_supply_chain, partial, pentest_only]
    ScanStatus:
      type: string
      enum: [queued, running, completed, failed, cancelled]
    JobStatus:
      type: string
      enum: [queued, running, succeeded, failed, skipped, cancelled]

    # ---------------------------------------------------------------- Auth
...
    CreateScanRequest:
      type: object
      description: >-
        `engines` is only read when `type` is `partial`. Note: unlike the
        `pentest_only` example in `documentation/07-api-specification.md`
        §5, the current implementation accepts no `target_id` field — this
        is a known doc/code gap, not something this file papers over.
      properties:
        type: { $ref: "#/components/schemas/ScanType" }
        engines:
          type: array
          items: { $ref: "#/components/schemas/EngineID" }
        branch:
          type: string
          description: Defaults to the repository's default branch when omitted/empty.
      required: [type]
    JobResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        engine: { $ref: "#/components/schemas/EngineID" }
        status: { $ref: "#/components/schemas/JobStatus" }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        finding_count: { type: integer }
        stats:
          description: Engine-specific stats shown on the UI's engine card (e.g. `files_scanned`, `rules_evaluated`).
          type: [object, "null"]
        error_reason: { type: [string, "null"], example: timeout }
        skip_reason: { type: [string, "null"], example: no_target_artifacts }
      required: [id, engine, status, started_at, finished_at, finding_count, stats, error_reason, skip_reason]
    ScanResponse:
      type: object
      description: >-
        Serves both `POST /projects/{id}/scans`'s 202 body and
        `GET /scans/{id}`'s 200 body. `risk` is always `null` in the
        current API version — risk scoring is not yet built.
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/ScanType" }
        status: { $ref: "#/components/schemas/ScanStatus" }
        requested_engines:
          type: array
          items: { $ref: "#/components/schemas/EngineID" }
        branch: { type: [string, "null"] }
        commit_sha: { type: [string, "null"] }
        queued_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        finding_counts:
          type: object
          description: 'Per-severity counts, e.g. `{"critical": 2, "high": 9}`.'
          additionalProperties: { type: integer }
        risk:
          description: Always `null` until risk scoring lands.
          type: "null"
        jobs:
          type: array
          items: { $ref: "#/components/schemas/JobResponse" }
        scan_number:
          type: integer
          description: 1-based position among the project's own scans (oldest = 1) — a human-readable "Scan #N".
      required:
        [id, project_id, type, status, requested_engines, branch, commit_sha, queued_at, started_at, finished_at, finding_counts, risk, jobs, scan_number]
    ScanSummaryResponse:
      type: object
      description: Deliberately lighter than ScanResponse — no per-job detail, used by the scan-history list.
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/ScanType" }
        status: { $ref: "#/components/schemas/ScanStatus" }
        branch: { type: [string, "null"] }
        queued_at: { type: string, format: date-time }
        started_at: { type: [string, "null"], format: date-time }
        finished_at: { type: [string, "null"], format: date-time }
        finding_counts:
          type: object
          additionalProperties: { type: integer }
        scan_number: { type: integer }
      required: [id, project_id, type, status, branch, queued_at, started_at, finished_at, finding_counts, scan_number]
    ScanListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/ScanSummaryResponse" }
        pagination: { $ref: "#/components/schemas/Pagination" }
      required: [data, pagination]
    OrgScanSummaryResponse:
      allOf:
        - $ref: "#/components/schemas/ScanSummaryResponse"
        - type: object
          properties:
            project_name: { type: string }
          required: [project_name]
    OrgScanListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/OrgScanSummaryResponse" }
        pagination: { $ref: "#/components/schemas/Pagination" }
      required: [data, pagination]
    EngineProgressResponse:
      type: object
      properties:
        engine: { $ref: "#/components/schemas/EngineID" }
        status: { $ref: "#/components/schemas/JobStatus" }
        progress_pct: { type: integer, minimum: 0, maximum: 100 }
        activity:
          type: string
          description: >-
            A real, live "what's happening right now" label — present only
            when the engine reported an actual named stage of its own
            (pentest's real phases, today). Absent for every other engine,
            which still gets a real (elapsed-time-based) progress_pct, just
            no stage name to go with it.
        finding_count: { type: integer }
      required: [engine, status, progress_pct, finding_count]
    ProgressResponse:
      type: object
      properties:
        scan_id: { type: string, format: uuid }
        status: { $ref: "#/components/schemas/ScanStatus" }
        progress_pct: { type: integer, minimum: 0, maximum: 100 }
        engines:
          type: array
          items: { $ref: "#/components/schemas/EngineProgressResponse" }
      required: [scan_id, status, progress_pct, engines]

    # ---------------------------------------------------------------- Findings
    LocationResponse:
      type: object
      description: >-
        A discriminated union keyed by `type` (`file`, `image`, `k8s`,
        `dependency`, `network`, ...) — the frontend switches on `type`;
        only the fields relevant to that type are populated, the rest are
        omitted (`omitempty` in the Go DTO, not present as `null`).
      properties:
        type: { type: string }
        path: { type: string }
        line_start: { type: integer }
        line_end: { type: integer }
        column: { type: integer }
        image: { type: string }
        layer_digest: { type: string }
        layer_index: { type: integer }
        file: { type: string }
        kind: { type: string }
        name: { type: string }
        namespace: { type: string }
        container: { type: string }
        field_path: { type: string }
        value: { type: string }
        from_helm: { type: boolean }
        chart_name: { type: string }
        template_file: { type: string }
        host: { type: string }
        ip: { type: string }
        port: { type: integer }
        protocol: { type: string }
        service: { type: string }
        url: { type: string }
        ecosystem: { type: string }
        package: { type: string }
        version: { type: string }
        manifest_path: { type: string }
      required: [type]
    EvidenceResponse:
      type: object
      description: >-
        Supporting evidence behind a finding's "More" view. `value` is
        always safe to render as-is — engines redact anything secret-shaped
        before a finding is ever emitted.
      properties:
        kind: { type: string, example: code_snippet }
        value: { type: string }
        redacted: { type: boolean }
        line_start: { type: integer }
        line_end: { type: integer }
      required: [kind, value, redacted]
    FindingListItemResponse:
      type: object
      description: >-
        Self-sufficient for a "More" detail expansion without a second
        request. `remediation` is deterministic, rule-authored guidance
        that must stand alone without AI. `first_seen_at`/`age_days` are
        not yet built (Phase 13, cross-scan history) and are absent here,
        not faked.
      properties:
        id: { type: string, format: uuid }
        engine: { $ref: "#/components/schemas/EngineID" }
        rule_id:
          type: string
          description: Permanent, namespaced `<engine>.<category>.<rule>` — appears in suppression comments, never renamed.
        title: { type: string }
        description: { type: string }
        remediation: { type: string }
        severity: { $ref: "#/components/schemas/Severity" }
        confidence: { $ref: "#/components/schemas/Confidence" }
        status: { $ref: "#/components/schemas/FindingStatus" }
        cwe:
          type: array
          items: { type: string, example: "CWE-89" }
        cve:
          type: array
          items: { type: string }
        owasp:
          type: array
          items: { type: string, example: "A03:2021" }
        cvss_score: { type: [number, "null"] }
        location: { $ref: "#/components/schemas/LocationResponse" }
        evidence:
          type: array
          items: { $ref: "#/components/schemas/EvidenceResponse" }
        source:
          $ref: "#/components/schemas/FindingSource"
        metadata:
          description: Engine-specific extra context (e.g. `impact`, `attack_path`, `osv_id`) — never load-bearing, absent keys are absent.
          type: object
      required:
        [id, engine, rule_id, title, description, remediation, severity, confidence, status, cwe, cve, owasp, cvss_score, location, evidence, source]
    FindingListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/FindingListItemResponse" }
        pagination: { $ref: "#/components/schemas/Pagination" }
      required: [data, pagination]

    # ------------------------------------------------------------------- Rules
...